[FIX] other pages not scrollable
